refactor(ui): #734 FileTreePanel の module-level state を zustand 化し責務分割
FileTreePanel.tsx (1091 行) が clipboard を module スコープの mutable
state で持ち、useState(0) + setter で強制 re-render するアンチパターンに
なっていた。clipboard 状態を zustand store (fileTreeClipboard) に移し、
ツリー描画責務を filetree/ サブモジュール (FileTreeChildren / FileTreeNode
/ FileTreeInlineRow / types / utils) に分割。FileTreePanel.tsx は 656 行に
縮小。振る舞いは変更していない純粋なリファクタ。

Closes #734
